tested on linux i686

+++ de.ibapl.jnhw.common/src/main/native/Callback_J_V_Impl.c
@@ -59,7 +59,7 @@ extern "C" {
     }
 
 #define TRAMPOLINE(index) \
-    void _jnhw_trampoline_J_V__ ## index (int value) {\
+    void _jnhw_trampoline_J_V__ ## index (int64_t value) {\
     JNIEnv *env;\
         if ((*jvm)->AttachCurrentThread(jvm, (void**) &env, NULL)) {\
             abort();\
