+++ de.ibapl.jnhw.common/src/test/native/Callback_J_V_Tests.c
@@ -38,7 +38,7 @@
 extern "C" {
 #endif
 
-    typedef void (*_callback_J_V)(int);
+    typedef void (*_callback_J_V)(int64_t);
     static _callback_J_V callbackPtr = NULL;
 
     /*
@@ -48,7 +48,7 @@ extern "C" {
      */
     JNIEXPORT void JNICALL Java_de_ibapl_jnhw_Callback_1J_1V_1Test_setCallback
     (JNIEnv *env, __attribute__ ((unused))jclass clazz, jobject callback) {
-        callbackPtr = UNWRAP_NATIVE_FUNCTION_POINTER_TO(void (*)(int), callback);
+        callbackPtr = UNWRAP_NATIVE_FUNCTION_POINTER_TO(void (*)(int64_t), callback);
     }
 
     /*
@@ -63,12 +63,12 @@ extern "C" {
 
 #ifdef HAVE_WINDOWS_H
     DWORD WINAPI thr_fn_J(LPVOID args) {
-            callbackPtr(*((int*)args));
+            callbackPtr(*((int64_t*)args));
             return 0;
     }
 #elif defined HAVE_PTHREAD_H
     void * thr_fn_J(void *args) {
-            callbackPtr(*((int*)args));
+            callbackPtr(*((int64_t*)args));
             return NULL;
     }
 #else
